+++ sdk/src/ecs/bridge/nativeEngineApi.generated.ts
@@ -67,7 +67,7 @@ export interface NativeEngineApi {
     gl_checkErrors?(context: string): number;
     gl_enableErrorCheck?(enabled: boolean): void;
     material_compileEsshader?(source: string, featuresCsv: string): number;
-    material_define?(materialId: number, shaderHandle: number, blendMode: number, flags: number): void;
+    material_define?(materialId: number, shaderHandle: number, blendMode: number, flags: number, depthBias: number): void;
     material_setTexture?(materialId: number, name: string, textureHandle: number): void;
     material_setUniform?(materialId: number, name: string, arity: number, v0: number, v1: number, v2: number, v3: number): void;
     material_undefine?(materialId: number): void;